Migrate org.apache.poi EBR to org.apache.poi BND
org.apache.santuario.xmlsec is no longer needed.

The previous 4.1.1 versions made all package imports optional; this does
the same, but it's not clear that this is correct.

Also add com.zaxxer which is a new dependency; it's not clear how
strictly this is needed, but the package import is optional.